3d Max pooling
1. max3d.cu : Performs max pooling over channels. Splits up input image
into tiles, each tile is a block that runs 16 threads to pool over the
channels of each of the 16 pixels belonging to this block/tile. This all
runs as part of 1 GRID. Note: I have created a device function called
getOffset_DEVICE that gets compiled for the device, we cannot have
common functions between device and host :/
2. utils.h : header file for (3).
3. utils.cu : Renamed from utils.c. Contains helper functions for
creating, filling and printing 3d/2d arrays. Also contains getOffset
(for functions on the host).